All of MoonEye's timer tests are now passing! This includes all timXX*.gb ROMs, along with tima_write_reloading.gb and tma_write_reloading.gb. Also fixed a regression in the IRQ blocking test, need to start working on regression testing ASAP.

+++ src/memory_map.rs
@@ -102,7 +102,7 @@ impl MemoryMap {
             StatTrigger(_) => vec![StatInt],
             _ => vec![]
         });
-        interrupts.append(&mut match self.timer.timer_cycle(cpu_cycles) {
+        interrupts.append(&mut match self.timer.timer_cycle(cpu_cycles as u16) {
             Some(_) => vec![TimerInt],
             None => vec![],
         });
